fix(engineering): vá SQL sai schema ở 3 route BIM/IoT chưa từng chạy
Sau khi vá tham số (commit trước), chạy thật trên Postgres lộ tiếp 4 lỗi schema:
- elements + simulate-4d dùng tasks.progress (cột thật: progress_percent)
- simulate-4d dùng tasks.approved_at (không tồn tại) — lấy mốc nghiệm thu sớm
  nhất từ task_history (status = 'nghiem_thu')
- simulate-4d lọc tasks.project_id (không tồn tại) — đi qua chuỗi
  work_packages → sheet_types → towers như lib/tien-do/report.ts
- iot/alerts PATCH: CASE WHEN $1 THEN $2 khiến Postgres suy $2 = text trong
  khi resolved_by là bigint — ép $2::bigint

Bộ quét db-params bỏ thêm comment mở đầu trước chuỗi SQL (trước đó mù với lời
gọi có comment chen giữa).

+++ app/api/engineering/iot/alerts/route.ts
@@ -82,7 +82,7 @@ export async function PATCH(req: Request) {
       `UPDATE engineering_iot_threshold_alerts
        SET is_resolved = $1,
            resolved_at = CASE WHEN $1 THEN NOW() ELSE NULL END,
-           resolved_by = CASE WHEN $1 THEN $2 ELSE NULL END
+           resolved_by = CASE WHEN $1 THEN $2::bigint ELSE NULL END
        WHERE id = $3 AND project_id = $4
        RETURNING *`,
       isResolved ?? true,
